New Updated Project
This commit is the pivot from being a project based on a specific
business to a project based in the general sector of the service
industry.

The code that is there has been commented, the login screen and the
register screen have been filled but aren't complete and the Parse API
has been implemented. It's a good starting point for everyone to use to
build upon.